modules/linux: return syscall number, not the key
Return the system call number as an integer value instead of
reinterpreting and returning the heap value pointer as an integer.
Otherwise the process is going to be killed due to a bad syscall error
since it's attempting to invoke system call number 200000 or something.

I committed this while I was developing a whole bunch of other stuff
at the same time and somehow it slipped past me when I tested it.
Maybe things were so buggy I didn't notice it.