elf: pack every ELF structure and union
This is necessary to make the offsets match the ELF specification.
My attempt to merge the 32 and 64 bit structures via union worked
fine in the ELF header structure but caused subtle issues due to
alignment and padding in the ELF segment structure since the common
part consists of just 4 bytes rather than 8. Also my liberal use of
nested structures didn't help matters.

Packing the structures and unions solves the offsets problem neatly
and it doesn't seem like there's going to be any alignment issues.

There is an unfixable problem though: the structure size is always
the size of the largest structure due to the union. Thus it is not
possible to perform simple pointer arithmetic when iterating over
the segments of a 32 bit ELF, the correct offset must be chosen
based on the input ELF's class or its header's segment size field.
A function that works around this problem by computing the pointer
manually will eventually be created.