embed: refactor segments structure
Keep both the bytes and the segments objects in the structure.
Useful for passing directly to functions as parameters.

Also use the new indexing function instead of iterating normally.
The non-standard layout of the structure violates array assumptions.

Also rename the program header table field to segment. Terse is better.