tools/embed: add ELF type conversion functions
Converts the pointers in the ELF embedder state structure
to a pointer to a lone ELF structure type.

Fixes gcc build errors because unsigned char * cannot be converted
implicitly to the structure types. Both GCC and clang reject this
code as invalid when compiled in my laptop and in GitHub actions.
The clang provided by Termux doesn't complain though. Weird...