lone: define bits functions
Add lone bits header and source files, lone's bitmaps/bitsets.
Implement general bit addressing functions as well as algorithms
for finding the first set or clear bit in the bit set.

This seems obscure at first but it's going to be surprisingly
useful in the planned refactoring of the memory allocator.
For example, used blocks could be tracked via bits which would
allow allocating memory by finding the first zero in the bitset,
flipping it and returning a pointer to the corresponding block.