make: add missing directories dependency
The generated NR.c and NR.list files as well as the intrinsic Linux
lone lisp module did not depend on the directories phony target
which sometimes causes issues in parallel builds because make can
attempt to create files inside the build tree before it exists.
This has been happening in the lone continuous integration builds,
causing them to fail for weird reasons.