workflows: use configuration for artifact upload
Access the build directory programmatically via the configuration
name set in the environment of the current workflow in the matrix.
No wildcards necessary.

Also prepend the name of the current configuration
to the name of the uploaded artifact. This enables
the upload of all artifacts generated by the entire
build matrix since they'll all have different names.

Apparently there's no way to use the CONFIGURATION
environment variable from the matrix in this context,
leading to incredibly annoying repetition of the logic
that makes up the configuration name. Can't be helped.