lone: implement generic stack data structure
Adds a useful data structure to lone: a general purpose stack
that can work generically with any data type. A necessary prelude
to the reification of the lone lisp stack and eventual implementation
of first class continuations.

Data is malleable and the operations being performed
on it determine its type. In other words, the stack
is a virgin memory block and there are lots of
functions for carving parts of it into something
meaningful and useful for the program.