types: use tagged pointers for lone values
This should cut down the size of register values to word size
down from a double word size representation, so more of them
will fit in processor caches.

The functions to access the parts of the value made it possible
to do this without major changes to the rest of the interpreter
and will also make it easier to make further adjustments to the
lone value representation.

The current implementation has a non-zero nil representation
which means zero filling vectors isn't enough. A second loop
has been added in order to set it to the proper value.
I should look for ways to eliminate that in the future.

Merry Christmas!