lisp: define lone lisp machine data structures
Define the data structures which will describe the lone lisp machine,
thereby laying the foundations for the implementation of a modified
version of the explicit control evaluator described in chapter 5.4
of the venerable Structure and Interpretation of Computer Programs.

All possible machine states are enumerated,
the structure of the stack frames is defined,
as are the registers of the machine itself.

References: https://mitp-content-server.mit.edu/books/content/sectbyfn/books_pres_0/6515/sicp.zip/full-text/book/book-Z-H-34.html#%_sec_5.4