lisp: define fundamental lisp machine functions
Define the functions which will reset and cycle the machine.

The machine's operation begins with a reset,
where all of its registers are set to neutral states
and where it is set up with an expression to evaluate.
It can then be cycled at the caller's discretion.
Each cycle causes the machine to perform the operations
as required by the structure of the provided lisp expression.
The machine will halt after it has reduced the expression
to a single value. It also returns control to the caller at nearly
all key points in the computation, making it easily interruptible
and resumable.