lisp: redefine primitive calling convention
Primitives will no longer be passed arguments lists directly.
Arguments will now be passed on the lisp machine stack.
Primitives will now be able to explicitly give up control
to the machine and be resumed later. To this end, primitives
must be able to save their data on the stack and restore it
at a later time. Passing arguments on the stack makes this
data passing mechanism regular and uniform.