lisp: implement machine cycle function
Evaluates the expression step by step, maintaining a stack as it goes
and returning control at key points. Atoms are returned as is while
lists are evaluated as applications of arguments to values.
Operators may be be lisp functions, primitive functions
or collections such as tables and vectors.
Arguments may or may not be evaluated.

This machine is a modified version of the explicit control evaluator
described in chapter 5.4 of the venerable Structure and Interpretation
of Computer Programs. It was manually translated from Scheme to C
and adapted to the unique needs of the lone lisp code base.

Notably, a resumption mechanism for primitives is implemented:
when they must evaluate more lisp code, they rig the machine
for expression evaluation and return an integer that represents
where they are supposed to resume from. The machine will evaluate
the expression and call them back with the resulting value and integer,
letting primitives resume work.

This works around the enmeshing stacks issue described by the developers
of Guile Scheme in a blog post by making C code interruptible and not
allowing the stacks to conflict with each other in the first place.

Unfortunately, making this work necessitated some code duplication.
The sequence evaluation cycles didn't work with primitives for reasons
that I couldn't fully understand. The problem was fixed by adding
a special sequence evaluation step just for primitives. The only
difference is a lisp machine register restore.

I hope I find a way to get rid of this ugly hack one day...

References: https://mitp-content-server.mit.edu/books/content/sectbyfn/books_pres_0/6515/sicp.zip/full-text/book/book-Z-H-34.html#%_sec_5.4
References: https://www.wingolog.org/archives/2010/02/26/guile-and-delimited-continuations