modules: rework primitives for lisp machine
Primitives no longer receive arguments and return values
via normal mechanisms, the lisp machine now passes and receives
arguments and return values to primitives via its own stack.

Primitives no longer call the evaluator in order to evaluate more code,
they set up the machine so it does what they want, save their state
on the lisp stack and return to the machine an integer that indicates
where in the function the execution should resume. This avoids problems
with enmeshing the C and lisp stack when continuations are implemented.

The implementation of primitives is now more complicated
but that's okay, the stack management power obtained
more than makes up for it and should enable
the implementation of delimited continuations.