modules/lone: add transfer primitive
Implement the transfer primitive which unwinds the lisp machine stack
back to the nearest continuation marker pushed by the control primitive,
copies all the stack frames into a continuation object
and calls the handler function given to the control primitive
with the specified value and the current continuation as arguments.
The return value of the handler function becomes the return value
of the control primitive, as if this function was returning from it.