lisp: capture control stack frames in transfer
Transfer now captures the control stack frames when it reifies
the current continuation. This allows it to be used in loops.

Without the control stack frames, resuming a continuation that
is looping over a collection and transferring values out of it
will crash the program on the next iteration.

    (control
      (each [10 20 30]
        (lambda (x)
          (print (transfer x))))
      (lambda (x continuation)
        (continuation (+ x 1))))

Currently, this prints 11 and then crashes.
This happens because transfer will look for the same control delimiter
it found in the first iteration but it will not be found because
the stack has already been unwound. Capturing the control stack frames
fixes the issue and makes the code work as expected.

Amazingly, it can be made to work by adding additional control frames
to supply the missing delimiters:

    (control
      (each [10 20 30]
        (lambda (x)
          (print (transfer x))))
      (lambda (x c1)
        (control
          (c1 (+ x 1))
          (lambda (y c2)
            (control
              (c2 (+ y 1))
              (lambda (z c3)
                (c3 (+ z 1))))))))

By capturing the control frames, this is no longer needed.