lisp/gc: extract cont. stack marker into helper
Extract the continuation stack marking algorithm into a helper function
and use to mark the general lisp stack in addition to the continuation's
saved stack frames. It's a precise algorithm that takes advantage of the
stack frame type information so it's better than using the native stack
marking logic on the lisp stack.